requirements.txt
mkdocs-material>=9.0
mkdocs-minify-plugin>=0.6
mkdocs-redirects>=1.2
pillow>=10.0
html5validator>=0.4
markdown-link-check>=0.2
pymdown-extensions>=10.0